requirements.txt
groq
python-dotenv
streamlit
GitPython 
PyGithub